Don't use the __OBJC__ compiler flag.
Being a header-only library, this causes crashes on cross objc++/c++ projects.
In particular, a logger created in c++ will call processFuncName (Util.h) on an objc++ function name
and segfault because the assumed characters are missing.

This particular fix attempts to do the non-objc++ processing on a function name,
but if this fails, defaults to the objc++ processing (passthrough).
The check is moved from compile-time to run-time.
This should eliminate the possibility of a crash.
It's unclear if there are certain function name formats that this might garble.